move synthetic_flight module from pycontrails/utis to pycontrails/ext
