+++ pycontrails/ext/synthetic_flight.py
