+++ tests/unit/test_cocip_grid.py
@@ -562,7 +562,7 @@ def syn_fl(instance_params: dict[str, Any], source: MetDataset) -> "SyntheticFli
         "level": source.data["level"].values,
         "time": np.array([t_start, t_stop]),
     }
-    SyntheticFlight = pytest.importorskip("pycontrails.utils.synthetic_flight").SyntheticFlight
+    SyntheticFlight = pytest.importorskip("pycontrails.ext.synthetic_flight").SyntheticFlight
 
     return SyntheticFlight(
         seed=5,
